--- pages/_app.tsx
@@ -1,12 +1,0 @@
-import type { AppProps } from "next/app";
-import { UserProvider } from "providers/UserProvider";
-import "styles/globals.css";
-
-function MyApp({ Component, pageProps }: AppProps) {
-  return (
-    <UserProvider>
-      <Component {...pageProps} />
-    </UserProvider>
-  );
-}
-export default MyApp;
