+++ src/app/layout.tsx
@@ -1,0 +1,20 @@
+'use client';
+import { UserProvider } from 'providers/UserProvider';
+
+import 'styles/globals.css';
+
+export default function RootLayout({
+  // Layouts must accept a children prop.
+  // This will be populated with nested layouts or pages
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  return (
+    <html lang="en">
+      <body>
+        <UserProvider>{children}</UserProvider>
+      </body>
+    </html>
+  );
+}
