+++ src/app/page.tsx
@@ -1,17 +1,9 @@
-import Head from "next/head";
-import Image from "next/image";
-import GoogleLoginProvider from "providers/GoogleLoginProvider";
-import styles from "styles/Home.module.css";
+import GoogleLoginProvider from 'providers/GoogleLoginProvider';
+import styles from 'styles/Home.module.css';
 
 export default function Home() {
   return (
     <div className={styles.container}>
-      <Head>
-        <title>Create Next App</title>
-        <meta name="description" content="Generated by create next app" />
-        <link rel="icon" href="/favicon.ico" />
-      </Head>
-
       <main className={styles.main}>
         <h1 className={styles.title}>shortcut master</h1>
 
@@ -23,10 +15,7 @@ export default function Home() {
             <p>このアプリについての説明</p>
           </a>
 
-          <a
-            href="https://github.com/vercel/next.js/tree/master/examples"
-            className={styles.card}
-          >
+          <a href="https://github.com/vercel/next.js/tree/master/examples" className={styles.card}>
             <h2>Examples &rarr;</h2>
             <p>実際の使い方を見てみる</p>
           </a>
@@ -42,9 +31,6 @@ export default function Home() {
           rel="noopener noreferrer"
         >
           Created by dak2
-          <span className={styles.logo}>
-            <Image src="/github.svg" alt="Github Logo" width={20} height={16} />
-          </span>
         </a>
       </footer>
     </div>
